requirements.txt
requests==2.32.3
gspread==6.1.2
google-auth==2.33.0
python-dateutil==2.9.0.post0
pytz==2024.1
